+++ dramatiq/middleware/__init__.py
@@ -1,4 +1,5 @@
 from .age_limit import AgeLimit  # noqa
+from .callbacks import Callbacks  # noqa
 from .middleware import Middleware, MiddlewareError, SkipMessage  # noqa
 from .prometheus import Prometheus  # noqa
 from .retries import Retries  # noqa